Edit linter errors
