--- capitalizeDecorator.rb
@@ -1,6 +1,0 @@
-require "baseDecorator"
-class CapitalizeDecorator <BaseDecorator
-    def correct_name
-      @nameable.correct_name.upcase
-    end
-  end
