+++ capitalizedecorator.rb
@@ -1,0 +1,6 @@
+require 'basedecorator'
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.upcase
+  end
+end
