+++ namable.rb
@@ -2,4 +2,4 @@ class Nameable
   def correct_name
     raise NotImplementedError, 'there is an error'
   end
-  end
+end
