+++ trimmerdecorator.rb
@@ -1,4 +1,4 @@
-require 'baseDecorator'
+require 'basedecorator'
 class TrimmerDecorator < BaseDecorator
   def correct_name
     @nameable.correct_name[0, 10]
